+++ nupy/src/neonutilities.egg-info/SOURCES.txt
@@ -3,15 +3,17 @@ README.md
 pyproject.toml
 src/neonutilities/__init__.py
 src/neonutilities/aop_download.py
-src/neonutilities/api_helpers.py
 src/neonutilities/citation.py
 src/neonutilities/read_table_neon.py
-src/neonutilities/zips_by_product.py
+src/neonutilities/tabular_download.py
 src/neonutilities.egg-info/PKG-INFO
 src/neonutilities.egg-info/SOURCES.txt
 src/neonutilities.egg-info/dependency_links.txt
 src/neonutilities.egg-info/requires.txt
 src/neonutilities.egg-info/top_level.txt
 src/neonutilities/__resources__/shared_flights.csv
+src/neonutilities/helper_mods/__init__.py
+src/neonutilities/helper_mods/api_helpers.py
 tests/test_aop_download.py
 tests/test_get_citation.py
+tests/test_zips_by_product.py
