+++ nupy/src/neonutilities/__init__.py
@@ -1,6 +1,4 @@
 from .citation import get_citation
-from .api_helpers import get_api
-from .api_helpers import get_zip_urls
-from .aop_download import by_file_aop
-from .aop_download import by_tile_aop
-from .zips_by_product import zips_by_product
+from .aop_download import by_file_aop, by_tile_aop
+from .tabular_download import zips_by_product
+from . import helper_mods
